Move uncertain_datetime functions
